Don't include requests exceptions in CanvasAPIErrors
Don't include exceptions from the `requests` library in the `"details"`
field of `CanvasAPIError`'s.

These details get sent to the client in the proxy API error response and
ultimately get shown to the user. `requests` library exceptions include
the requested URL in their `args` and as part of their string
representation. With OAuth 2 APIs such as the Canvas API the requested
URL sometimes contains sensitive strings that should never leave the
server, such as refresh tokens and client secrets.