Rename CanvasOAuth2RedirectResource to OAuth2RedirectResource
There's nothing Canvas-specific about this. It only has the
Canvas-specific name because it's only used in Canvas currently, but I'm
pretty sure we're gonna end up using this for Blackboard OAuth 2
redirects as well.

The only reason this view exists is because the
`"canvas_oauth_callback"` and `"canvas_api.authorize"` routes that use
it (the two routes for the authorization popup window) need
`context.js_config` to be set. They can't use the default resource
because it doesn't have `context.js_config`. And `context.js_config`
can't be added to the default resource because that resource is used by
a bunch of views that don't and can't have `context.js_config`.