Fix up some inconsistent route names
Rename the routes `"canvas_api.authorize"` and `"canvas_oauth_callback"`
to `"canvas_api.oauth.authorize"` and `"canvas_api.oauth.callback"` to
be consistent with the other `"canvas_api.*"` routes and to use the
`"."` nicely for grouping / namespacing the routes.

Also fix up the Canvas API authorize URL from `"/api/canvas/authorize"`
to `"/api/canvas/oauth/authorize"`, grouping auth-related things under
`"oauth"`, the same as the route names do.

We can't easily change the actual path associated with the OAuth
callback route unfortunately because that path has already been
conigured in lots of LMS instances. I think we could actually do it but
we would have to update our install documentation and we'd have to
maintain backwards-compat support for the "legacy" URLs. This would be
doable but I'm not going to bother with it.