Remove an unnecessary test
This test isn't adding anything -- the code in `routes.py` is entirely
declarative so there's not much point in testing it. You could test that
it at least doesn't crash given a realy Pyramid config object, but this
test uses a mock so it doesn't even do that. And in fact `routes.py` is
already being executed as part of the `pyramid_config` fixture, so that
is already tested:

https://github.com/hypothesis/lms/blob/d4ddd02113ae4e4e544eb41897106bb4be0293f6/tests/unit/conftest.py#L105

Hence why we still have 100% coverage after removing this test.