Isolate TokenStore in unit tests
Make it so that the only unit tests that instantiate a `TokenStore` object
are the ones in `_token_store_test.py`. All others use a an autospecced
mock `TokenStore`.

This is so that `TokenStore` can be moved from
`lms.services.canvas_api._token_store.TokenStore` to a top-level service
(`lms.services.token_store.TokenStore`) so that it can be shared by
`CanvasAPIClient`, the upcoming `BlackboardAPIClient`, and possibly
others in future.

When `TokenStore` is a top-level service it no longer makes sense for
unit tests to consider it a "close collaborator" of `CanvasAPIClient'
and not mock it. `TokenStore` has a very simple API: it has only
`save(access_token, refresh_token, expires_in)` and `get()` (no
arguments) methods so it's very simple to mock and to assert that users
like `CanvasAPIClient` and `BlackboardAPIClient` use it correctly.

The test isolation will also make the actual change to move `TokenStore`
to a top-level service easier to make: the resulting test failures will
be easier to understand and the tests will be easier to update.