Fix a bad patch fixture
This patch just wants to create a mock `TokenStore` object so that it
can be passed as an argument to `AuthenticatedClient.__init__()` (this
is a little difficult to track down: the mock `token_store` is passed to
`AuthenticatedClient.__init__()` in the `authenticated_client` fixture
`tests/unit/lms/services/canvas_api/conftest.py`). The test **does not**
want to replace the name
`"lms.services.canvas_api._token_store.TokenStore"` (in
`_token_store.py`'s namespace) with a mock. The use of `patch()`, which
replaces names in namespaces, is therefore incorrect and misleading.
`create_autospec()` is the right tool to use here.

`patch()` also returns a mock _of the `TokenStore` class_, not a mock of
an object instance of the class which is what's actually wanted here.
There are all sorts of differences between mocks-of-classes and
mocks-of-instances, for example mocks-of-classes are callable (and
return a mock instance of the class!). The tests were getting away with
this because they happened not to trip on any of the differences but it
was incorrect and could have led to problems (either test failures, or
tests passing that should fail). The `instance=True` argument to
`create_autospec()` makes it return a mock instance not a class.

The unit tests for a `foo.py` module should only ever path names within
that `foo.py` module's namespace. If `foo.py`'s unit tests patch a name
within `bar.py` (which `foo.py` imports) the patch won't actually apply
in time, `foo.py`'s will end up using the real object not the mock
because `foo.py` will have imported the name from `bar.py` before the
patch gets executed. These tests aren't running into this problem
because they're not actually relying on the patch, they're passing the
mock `token_store` to `AuthenticatedClient.__init__()` as an argument.

See this GitHub comment for another example and detailed explanation of
the issue with misusing `patch()` in this way:

https://github.com/hypothesis/via3/pull/222#discussion_r469865055