Make canvas_api._token_store a top-level service
Move `services.canvas_api._token_store.TokenStore` to a top-level service
(`services.token_store.TokenStore`) and register it with
pyramid_services as the `"token_store"` service.

This is because `TokenStore` isn't actually Canvas-specific: it's just a
get-and-set service for `models.OAuth2Token`, which is a generic model
for storing OAuth 2 tokens. `TokenStore` is used by `CanvasAPIClient`
and will soon be used by the upcoming `BlackboardAPIClient` and possibly
others in future.