Rename AuthenticatedClient._token_store to oauth2_token_service
I think it's a little weird for the name of the `"oauth2_token"` service
to change to something completely different (`token_store`) when it gets
passed to `AuthenticatedClient`. Makes the code hard to follow / makes
it hard to know what `token_store` actually is. So rename `token_store`
to `oauth2_token_service`. This is consistent with how references to
other services are named, elsewhere in the code