Don't raise CanvasAPIAccessTokenError from OAuth2TokenService
It doesn't make sense to raise a Canvas-specific error from this generic
OAuth 2 service. Add a new `NoOAuth2Token` exception and raise that
instead.

Catch `NoOAuth2Token` in the (Canvas-specific) `AuthenticatedClient`
that calls `OAuth2TokenService` and wrap it in
`CanvasAPIAccessTokenError`, so the rest of the Canvas API services and
views are unaffected