Add fake Blackboard API views
Add the minimum fake Blackboard API views necessary to enable the
frontend code for the Blackboard Files feature to be developed:

`views.api.blackboard.authorize.authorize()` is the view that the
frontend will open in a popup window to get an OAuth 2 access token from
Blackboard. In future this will redirect the browser to Blackboard's
authorization page, which will then redirect the browser to our
`oauth2_redirect()` view after the user clicks **Authorize**. For now,
our `authorize()` view just redirects directly to our
`oauth2_redirect()` view, skipping Blackboard.

`views.api.blackboard.authorize.oauth2_redirect()` is the view that
Blackboard will redirect the browser to with an authorization code that
the server can exchange for an access token. For now it just saves a
fake access token to the DB and returns.

The user will see the popup window close immediately, without actually
taking them to the Blackboard authorization page.

Once the real versions of these authorization views are written they
might be merged with the corresponding Canvas views, rather than having
separate Blackboard and Canvas views.

`views.api.blackboard.files.list_files()` returns a hard-coded list of
Blackboard files for the frontend to display.

`views.api.blackboard.files.via_url()` returns different hard-coded
public PDF URLs depending on which of the hard-coded files from
`list_files()` was selected when creating the assignment.

These views also both check that `oauth2_redirect()` has added a fake
Blackboard OAuth 2 token for the user to the DB. If not they raise
`NoOAuth2Token` which will cause the frontend to show an **Authorize**
button that will open the `authorize()` view in a popup window when
clicked, beginning the process of adding the fake token to the DB.