Suppress DeprecationWarning from pyramid.path
In order for this suppression to work, the `-Werror` flag was removed.
Non-ignored warnings will still cause tests to fail due to the
configuration in pytest.ini. The current combination of pytest flags and
pytest warning filters matches the configuration in h.

This warning is not new but is triggered as a result of upgrading pytest
to v6.1.0. See https://github.com/hypothesis/h/pull/6273/commits/be60099419a1a45b9ec6759576b4320bc4203be5.