Display a warning banner in old + unsupported browsers
Add a new `browser_check` bundle which is loaded before the main
`frontend_apps` bundle. This bundle is compiled to work in older
browsers and it performs feature tests to check whether the current
browser is likely to work. If the check fails it displays an advisory
banner at the top of the page.

When the banner is shown this does not stop the frontend app from
attempting to load. That way we can accept rare false positives in these
checks and still let users attempt to use the LMS application.

The `async` attribute was removed from the frontend bundle scripts to
ensure that the scripts are executed in-order. It is important that the
browser_check script executes before the frontend_apps script which
might attempt to load polyfills.