Fix content iframe size in grader in Firefox
The content `<iframe>`'s size was set using a `height="100%` attribute.

In situations when the grader bar is shown, this caused the iframe to be
the wrong height in Firefox (but not Chrome). Instead of being the
height of the viewport minus the height of the grader, it ended up being
the same height as the viewport.

This PR fixes the issue by using `flex-grow: 1` to make the iframe fill
the remaining space in the `LMSGrader` parent after the grader is
allocated space. This produces the same results in all browsers.

To make this sizing method work in contexts where the `LMSGrader`
is not used, the `BasicLtiLaunchApp__content` container was made to have
the same `display: flex; flex-direction: column` layout as `LMSGrader`.
This required also specifying the height of the
`BasicLtiLaunchApp`'s root `<div>`.

This commit also removes an unused `js-via-iframe` CSS class.

Fixes https://github.com/hypothesis/support/issues/149