Show custom error dialog for Canvas file permission errors
When the backend reports that fetching the content URL for an assignment
launch fails due to a Canvas file permission error, indicated by
`"error_code": "canvas_api_permission_error"` in the API response, show
a specific dialog for that error. This dialog explains common causes of
the problem and how to rectify them. It also differs from the generic
error dialog in that clicking "Try again" does not retry authorization,
since we know that is not necessary.