Make a pytest fixture more flexible
The method() fixture currently assumes that add_canvas_file_id() and
add_document_url() can be called with the same arguments. Future changes
to add_canvas_file_id() are going to add more required arguments to it
so this fixture will no longer work. Change the fixture to be capable of
calling each method with different args.