Add AssignmentService
Add an `AssignmentService` for getting and setting assignment document
URLs, instead of using classmethods on the model.

The main reason for doing this is that it makes the code more efficient
because the service can cache SQL query results on a per-request basis
(due to the design of pyramid_services: it lazily instantiates one
singleton instance of each service _per request_) and avoid repeated DB
queries for the same `ModuleItemConfiguration` within a single request.

Different parts of the code can blindly call the service's methods
without regard to whether another part of the code might have called the
same method in the same request, and the service-layer caching will
ensure that the DB queries aren't repeated.

This also simplifies the interface because `get_document_url()` and
`set_document_url()` don't need to have the `db` param passed in to
them. It can be injected by the service factory.