Don't show error for "insufficient scopes"
If we get an `"Insufficient scopes on access token"` error from the
Canvas API just show the user the non-error *Authorize* dialog so they
can get us a new access token (we'll always request one with the needed
scopes). Don't show the user an error dialog: this is an expected /
unavoidable circumstance, and getting a new access token should always
fix the problem.

Fixes https://github.com/hypothesis/lms/issues/2345

Unfortunately we have no choice but to use the English-language string
`"Insufficient scopes on access token."` to identify this particular
type of error. Canvas doesn't provide any kind of error codes meant for
machines. The HTTP status can't be used because the same status is
shared by different types of error that we need to respond differently
to.

I've tested in three languages and `"Insufficient scopes on access
token."` doesn't appear to get localized into the user's account's
language as many of the strings in Canvas API responses do. As with
other Canvas API response strings that we have no choice but to rely on,
we have no way of knowing for sure that Instructure won't localize it in
the future. If that happens, they'll break our app.

The string doesn't appear to be internationalized in Canvas's source
code:

https://github.com/instructure/canvas-lms/blob/20f91713bc3b0524bf1a3bc580d9b6fa51d6053a/app/controllers/application_controller.rb#L1644-L1645