Support VitalSource bookshelf launches in the LMS frontend
This commit adds support in the frontend for loading the VitalSource book viewer
in an LTI launch, when the backend provides the necessary configuration.
The backend logic needed to generate the form parameters is not yet
present and will be added separately.

Changes in this commit:

 - Check for VitalSource configuration from the backend in
   `BasicLtiLaunchApp` and respond by rendering a
   `VitalSourceBookViewer` instead of the normal Via iframe
 - Add a `VitalSourceBookViewer` component which creates an empty
   iframe, renders form parameters from the backend into it and
   auto-submits the form.

See https://github.com/hypothesis/lms/pull/2390 for an overview of the
VitalSource launch process.