Fix `configured` predicate test in LTI launch views
The `configured` predicate incorrectly evaluated to `False` for
URL-configured assignments because the check for the assignment being
URL-configured was accidentally removed in a08b67464c48479290febf4ed6f04cb653c652c3.

The issue went unnoticed because the tests for the `Configured`
predicate had a flaw that assignments where DB-configured "by default"
and so the URL-configured tests passed when they should not have. This
commit fixes that by making assignments not DB-configured.