Refresh client's grant token if it has expired
During an LTI launch the LMS backend generates a grant token for use by the
Hypothesis client that is valid for 5 minutes (see `JSConfig._grant_token`).

If the Hypothesis client attempts to fetch its configuration after this
time has elapsed then the client will fail to log in. With the sidebar
application this rare because the sidebar launches as soon as Via loads.
It could happen however if the LMS frontend did not load the assignment
content immediately but instead paused for Canvas OAuth and the user
took more than 5 minutes to respond. With the client's "Notebook"
application this is expected to be much more common however, because the
notebook isn't loaded until the user explicitly chooses the "Open
notebook" menu option.

Fix the issue by adding a new backend API view, `/api/grant_token` that returns
a new grant token and calling it from the LMS frontend if needed before
responding to a request from the Hypothesis client (which could be from
the sidebar or notebook) for configuration.

 - Add `/api/grant_token` route in `lms.views.api.grant_token`
 - Modify `ClientRpc` to check expiry time of grant token and refresh it
   if needed before responding to a `requestConfig` RPC call from the
   client

Part of https://github.com/hypothesis/client/issues/2801