Refresh client's grant token if it has expired
During an LTI launch the LMS backend generates a grant token for use by the
Hypothesis client that is valid for 5 minutes.

If the Hypothesis client attempts to fetch its configuration after this
time has elapsed then the client will fail to log in. With the sidebar
application this rare because the sidebar launches as soon as Via loads.
It could happen however if the LMS frontend did not load the assignment
content immediately but instead paused for Canvas OAuth and the user
took more than 5 minutes to respond. With the client's "Notebook"
application this is expected to be much more common however, because the
notebook isn't loaded until the user explicitly chooses the "Open
notebook" menu option.

Fix the issue by modifying the `ClientRpc` class in the LMS frontend to
request a new grant token if the current one has expired when responding
to a `requestConfig` RPC request from the Hypothesis client.

 - Add a `JWT` helper class that wraps a server-generated JWT and
   provides methods to check whether it is still valid.

 - Modify `ClientRpc` to use the `JWT` helper to check if the JWT is
   still valid before returning config to the client, and fetch a new
   grant token otherwise.

Part of https://github.com/hypothesis/client/issues/2801