Add Issued At (iat) timestamp to grant tokens
This contains the same information as `nbf` (Not Before) but is
explicitly the time at which the token was issued, whereas `nbf` might
(but currently doesn't) incorporate some leeway.

The `iat` timestamp is going to be used by the LMS front end to determine
the skew of client and server clocks.