Render more details with LTI outcome errors
In order to help pinpoint moodle grading errors, it will help to relay more information to the client so that it may be captured in a screen shot at the time the error occurs. This helps prevent losing metadata when the logs expire in paper trail after too much time.