Use `pycryptodomex` instead of `pycryptodome`
The 'x' varient changes the name of the installed package so it's no
longer a direct replacement for `pycrypto`. This means it can't conflict
with it and both can be installed. It _also_ means that tools which
inspect the virtual env can see it's not installed and don't get confused
between the two.