Add ProxyAPIError and ProxyAPIAccessTokenError
This is some ground work before adding in Blackboard API support.

A new `ProxyAPIError` base class is added which both `CanvasAPIError`
and `LTIOutcomesAPIError` inherit from. In future a new
`BlackboardAPIError` will also inherit from this. The
`proxy_api_error()` exception view that was previously associated with
`CanvasAPIError` and `LTIOutcomesAPIError` is now associated with
`ProxyAPIError` so if future Blackboard code uses `ProxyAPIError` it'll
get the exception view behavior for free.

`CanvasAPIAccessTokenError` is renamed to `ProxyAPIError` so that it can
be reused by the Blackboard API code. For example
`CanvasAPIAccessTokenError`'s are filtered out from Sentry. Since that's
now been changed to a `ProxyAPIAccessTokenError` that the Blackboard
code can also use, the Blackboard code will get the Sentry filtering for
free. There was nothing actually Canvas-specific about
`CanvasAPIAccessTokenError`.