Changed wording of button in LMSFilePicker
On [this commit](https://github.com/hypothesis/lms/commit/c7cc79f78241030bbb8b7791aa943b3b1bc3934a)
the wording of the retrying authorization button was changed to
`Authorize again`. There has been a
[discussion about that wording](https://app.slack.com/client/T03QZM0HN/C1M8NH76X/thread/C1M8NH76X-1620218640.044700)
and it has been agreed that it is better to revert it back to `Try again`.