Raise ProxyAPIAccessTokenError not NoOAuth2Token
You would need the frontend changes from
https://github.com/hypothesis/lms/pull/2636 to actually test this, but:

`ProxyAPIAccessTokenError` is the one that has an exception view:

https://github.com/hypothesis/lms/blob/e174b161e90918c94402b8e662e150b063b67b42/lms/views/api/exceptions.py#L107-L109

The exception view returns a response that triggers the frontend to show
a "You need to authorize" dialog.

If the view raises `NoOAuth2Token` instead that gets caught by the
catch-all exception view that triggers a "Something went wrong" dialog.