Refactor selected content representation in `FilePickerApp`
Combine the various state fields indicating the type of content selected
for an assignment into a tagged union. This is a precursor to extracting
the content selection UI into a separate component. It also better
represents the fact that only one type of content can be selected for an
assignment.

This would have required a change to the hidden `document_url` field.
This field does not appear to be used for anything and is not a
field that appears in the specs for the Content-Item form [1], so I have
removed it.

[1] https://www.imsglobal.org/specs/lticiv1p0/specification