Merge `contentItemFor*` functions
Combine the `contentItemFor<ContenType>` functions into a single
`contentItemForContent` function which accepts a `Content` object as an
argument, as well as additional query params to add to the launch URL.

Combining these functions removes the need to switch on the content type in other
places in the code.

The support for adding additional query params will be used to add the
`group_set` query param in assignments that are configured to use
groups.