Refactor internal state of the Canvas file picker dialog
While trying to make some changes to the way the "fetching" state is
rendered in order to be consistent with upcoming UI for the VitalSource
book picker I found it confusing to understand the possible states of
the dialog and the rendered output in each.

This commit refactors the internal states of the `LMSFilePicker` component to
try and simplify the situation.

 - The single `DialogState` type has been converted into a union that
   specifies the 4 main states: Prompting for authorization,
   fetching files, showing a file list, displaying an error

 - A union has been introduced to describe the 3 possible actions that
   the dialog may offer as the main affirmative action for the dialog:
   Authorize, Reload, Select a file

 - Rendering of messages associated with the "authorizing" and "error"
   states has been inlined into the main UI rendering at the end of
   `LMSFilePicker`

 - Switching on union discriminator fields has been converted to use
   `switch` statements for clarity, per [1]

[1] https://hypothes-is.slack.com/archives/C1M8NH76X/p1621942239121100