Deduplicate `requests` usage
Problem
-------

There's a bit of logic that's repeated in a few places about using
`requests` to send a request, receive the response, and validate the
response using a `marshmallow` schema, with associated error handling
and tests, re-using a TCP session, making sure appropriate timeouts are
always used, etc. This is a basic version of the logic that's repeated:

```python
session = requests.Session()

try:
    response = session.request(method, url, timeout=9, ...)
    response.raise_for_status()
except requests.RequestException as err:
    raise CustomHTTPErrorClass(response) from err

validated_data = None

if schema:
    try:
        validated_data = schema(response).parse()
    except lms.validation.ValidationError as err:
        raise CustomHTTPValidationErrorClass(response) from err

return (response, validated_data)
```

This logic (or very similar) is repeated in `HAPIClient`,
`LTIOutcomesService` and `CanvasAPIClient` (in the `BasicClient` helper
class):

* https://github.com/hypothesis/lms/blob/b995d470acf28f70c342b03d0d4acd45b4c508c1/lms/services/h_api.py#L83-L99
* https://github.com/hypothesis/lms/blob/b995d470acf28f70c342b03d0d4acd45b4c508c1/lms/services/lti_outcomes.py#L101-L119
* https://github.com/hypothesis/lms/blob/b995d470acf28f70c342b03d0d4acd45b4c508c1/lms/services/canvas_api/_basic.py#L85-L95

The upcoming `BlackboardAPIClient` also needs to send HTTP requests
using `requests` and validate the responses using `marshmallow` in
exactly the same way, but that would mean duplicating this similar code
and tests for a fourth time.

Solution
--------

Add a new `HTTPService` to do the send-request-and-validate-response
bit. This will be implemented once, correctly, in `HTTPService` and
thoroughly unit tested there. Other services will just need to use a
very simple interface that `HTTPService` will provide, and their unit
tests will just need to work with a mock of that very simple interface.
A large amount of code and tests will be simplified and duplication
removed. Our code's behaviour (e.g. robustness) will be improved because
send-request-and-validate-response will be implemented and testedonce
carefully and correctly, instead of several times with varying degrees
of correctness.

This PR adds the `HTTPService` and refactors `LTIOutcomesClient` and
`HAPI` and their tests to use `HTTPService`.

The new `BlackboardAPIClient` that'll use `HTTPService` will come in a
later PR.

Refactoring `CanvasAPIClient` to use `HTTPService` will also come
separately: that ones a bit more work.