Update `FilePickerApp` layout to match mocks
Update the `FilePickerApp` layout to match the layout in [1]. This
layout consists of a two-column grid with an "Assignment details" header
and a footer containing buttons to navigate between steps of the dialog.
Currently the footer contains only a "Continue" button which is shown
when the group configuration step is active. The left column shows the
current step and the right column contains the content for that step.

[1] https://github.com/hypothesis/lms/issues/2523#issuecomment-821160185.