Shorten HTTPService's docstring
I had thought that pulling out several of `requests`'s most useful
parameters and documenting them here would be a useful cheat-sheet. All
of this information is in the requests docs but not necessarily easy to
find or all in one place in those docs.

But the really long docstring is just creating too much space between
the method header and body and all the documented `requests` parameters
are crowding out our own custom parameters (such as `schema`, and more
in future) which actually makes the docstring less useful.