Truncate long URLs in "Assignment details" view
When displaying a short description of the content for an assignment,
shorten long URLs so that the form does not overflow the viewport.

We could have gotten away with some very simple text truncation here,
but I've tried to do a little better and truncate the URL in a way that
preserves the most informative parts. I suspect we'll have use for this
logic in other places.

Fixes #2774