Show saving indicator when assignment creation form is submitted
Since the assignment creation form is shown in an iframe, there is no
indicator in the browser when the form is being submitted and the
browser is waiting for the server's response. This is not a major
problem because this typically happens quickly, but in case it is slow
show a spinner to indicate that something is happening.