Add gap between rows in Assignment details form
Fix an issue where the "Assignment content" and "Group assignment" rows
of the "Assignment details" form could touch visually if the assignment
content row flowed onto multiple lines. The left column had a bottom margin
set but the right column did not.

Fix the issue by replacing the left column's bottom margin with a
`row-gap` which applies to the whole row. This gap is additive to any
cell margisn, so the bottom margin for the header row also had to be adjusted.