Add space between checkbox and group set select
Wrap the contents of the space below the "This is a group assignment"
checkbox in a `GroupConfigSelector__select` div with a top margin. The
top margin is the same as that of a `<p>` element so the visual margin
is the same whether this area is displaying the authorization prompt or
the group set `<select>` control.