Don't create_all when alembic is present on make db
If starting from a empty DB we should create_all and stamp the DB with
the correct alembic version so migrate head does the right thing.

If using `make db` on a non empty DB we should rely on alembic to get us
to the right state, ie don't `create_all` or stamp the DB if alembic is
present.