Add infrastructure for service layer in frontend
Add a `Services` context provider and associated `useService` hook which
can be used to expose and access stateful services to components in the
tree.

Services are registered and retrieved using the class as a key. Service
instances do not have to be instances of the class, although they
typically will be, but they must implement the same interface. Using the
class as a key allows TypeScript to infer the return type of
`useService`.