Convert grade fetching and submission to use a service
Replace the API wrapper functions in `utils/grader-service` with a
`GradingService` service and change the `SubmitGradeForm` component to
use it.

At present this change only has a small benefit in that it avoids
`SubmitGradeForm` needing to know about how to extract API auth
information from the `Config` context and pass it to `apiCall`. As the
application grows though the service layer provides a place to put
stateful non-UI logic, keeping UI components simpler and easier to
reason about and test.