Add resource_link_id param to via_url API
Pass the assignments `resource_link_id` (assignment ID) to the Canvas
Files `/via_url` API.

In future PRs the `/via_url` API is going to need the `resource_link_id`
to fix course-copied Canvas files assignments. It's needed to look up
the assignment's `ModulteItemConfiguration` (assignment configuration)
in the DB in order to get and set "Canvas file mappings" in the
`ModulteItemConfiguration.extra` column.