Add CanvasService.find_matching_file_in_course()
A future PR is going to add business logic to
`CanvasService.public_url_for_file()` that will call this
`CanvasService.find_matching_file_in_course()`.

I've made `find_matching_file_in_course()` a separate public method
because it's a logically separate unit that *could* be used by other
code outside of `CanvasService` even though it isn't going to be right
now, and because having it be a separate public method encourages a good
method name, good docstring, and its own separate unit tests. Small
separate units instead of one big unit where a top-level public method
calls lots of private methods.

When searching for a matching file in a course we do this by calling the
Canvas API to get a list of all the files in the course, rather than
looking in our `models.File` table in the DB, because:

* It always gets the most up to date list of files from the course,
  doesn't return false positives or false negatives
* It only returns the files that the current user can see, which is what
  we want. For example students can't see "unpublished" files but
  teachers can. Our `models.File` table doesn't know about this