Check cache buster when serving static asset requests
If a static asset was requested with a cache buster query string which
did not match the manifest (in `build/manifest.json`) this could lead to
a downstream proxy (eg. Cloudflare or the user's browser) storing the
returned asset under the wrong cache key.

A scenario where this could happen is if the application was served from
multiple hosts and a new version is deployed in a rolling fashion. In
this scenario if an HTML page is served from version X+1 of the
application and the browser loads an asset referenced by the page, which
hits an app server hosting version X of the application, then the
browser or Cloudflare could end up storing version X of the asset under
the URL for version X + 1. This would lead to subsequent requests for
version X + 1 of the asset returning the old value until the cache was
cleared.

To prevent this problem entirely we either need to:

 1. Deploy the web app in such a way that an HTML page and associated asset requests
    cannot hit different versions of the application

    OR

 2. Separate the deployment of static assets from the deployment of the
    application (eg. by deploying assets to a CDN prior to application
    deployment)

This PR reduces the impact of the issue by verifying the cache buster
query string used in an asset request and returning a 404 response if
incorrect, which should prevent or lessen downstream caching.