Remove CORS header from static asset responses
This code was copied from the "h" project where it used to be necessary
for the Hypothesis client. It was never needed in the lms context
because static assets are only loaded from pages served from the lms's
origin, and is no longer needed for h because the Hypothesis client is
served from a separate domain.