Move "marshalling" code from schemas to views
Make the BlackboardAPIClient service (and its schemas) just return
Blackboard API dicts in the format given by the Blackboard API, without
modifying them.

The service *does* remove unwanted fields that we don't care about from
the dicts. In the future we'll actually add Blackboard API query params
to request only the fields we want and save network bytes.

The service does also validate that the Blackboard dicts are as we
expect.

But it doesn't mutate them any further. Instead it's the view's job to
not about our proxy API's response format and transform the dicts into
that format.