Move BasicBlackboardAPIClient into BlackboardAPIClient
Rename the BasicBlackboardAPIClient class to just BasicClient and move
it into BlackboardAPIClient. It's now accessible as
BlackboardAPIClient.api and is no longer a top-level service.

Note that BasicClient is still public and is still used (via
BlackboardAPIClient.api) by some external code. In a future PR it'll
become completely private.