Search by both original and mapped file ID
When launching a Canvas Files assignment and the file ID isn't found in
the course, use both the assignment's original file ID and its mapped
file ID (if any) to search for a matching file in the course.

I think this might fix an obscure Canvas course copy edge case involving
renaming files in Canvas and possibly requiring chains of multiple
course copies. And it shouldn't do any harm.