Move access_token_request() into OAuthHTTPService
Move the Blackboard API client's BasicClient.get_token() to
OAuthHTTPService.access_token_request().

This is mostly because I want a future commit to add
refresh_token_request() to OAuthHTTPService as well, and
access_token_request() and refresh_token_request() go together (they
even share code).

The method was renamed from get_token() to access_token_request()
because I think access_token_request() is a better name as it sends an
"access token request" (what the OAuth 2.0 spec calls this particular
request), and the name goes nicely with refresh_token_request() which
will be added in a future commit (and which will send what the spec
calls a "refresh token request") and with the existing request().

I'm not sure that this access_token_request() method is fully generic
yet. It sends the type of access token requests that the Blackboard API
requires. In particular it sends the client ID and secret in a basic
HTTP auth header. I think some servers may require these in the body. We
can adapt the method as needed as we add more servers.