Update LMSFilePicker to filter out folders from BB files API results
This is an interim change that makes the front-end file-picker for
BlackBoard files function on top of changes to the proxy API to
retrieve a list of files. The returned results now contains some extra
properties, and include both files and folders.

For the moment, make `LMSFilePicker` filter out folders such that the
interface behaves as it did before the API change: present only root-level
files.